@@ -38,13 +38,11 @@ public class OpenIconCommand extends IconCommand {
 			 * Delay the task, since this command is executed in ClickInventoryEvent
 			 * and opening another inventory in the same moment is not a good idea.
 			 */
-			Bukkit.getScheduler().scheduleSyncDelayedTask(ChestCommands.getInstance(), new Runnable() {
-				public void run() {
-					if (player.hasPermission(menu.getPermission())) {
-						menu.open(player);
-					} else {
-						menu.sendNoPermissionMessage(player);
-					}
+			Bukkit.getScheduler().scheduleSyncDelayedTask(ChestCommands.getInstance(), () -> {
+				if (player.hasPermission(menu.getPermission())) {
+					menu.open(player);
+				} else {
+					menu.sendNoPermissionMessage(player);
 				}
 			});
 
