@@ -47,7 +47,7 @@ public class SpecialConfig {
 
 		// Check if the configuration was initialized
 		if (defaultValuesMap == null) {
-			defaultValuesMap = new HashMap<String, Object>();
+			defaultValuesMap = new HashMap<>();
 
 			// Put the values in the default values map
 			for (Field field : getClass().getDeclaredFields()) {
